T8.3 - Implement Tables and Dialogs
Refactor GUI components to improve code organization and maintainability:

- Extract ProjectsTable from projects_tab.py to widgets/projects_table.py
- Extract CategoriesTable from categories_tab.py to widgets/categories_table.py
- Extract AddProjectDialog and EditProjectDialog to dialogs/add_project_dialog.py
- Extract AddCategoryDialog and EditCategoryDialog to dialogs/add_category_dialog.py
- Update import statements in tab files and test files
- Maintain all existing functionality and test coverage

This refactoring improves:
- Separation of concerns (each file has single responsibility)
- Testability (components can be tested in isolation)
- Reusability (tables and dialogs can be used elsewhere)
- Code organization (follows established widget/dialog directory pattern)

All 56 tests pass (26 projects + 30 categories tests)
