+++ src/ocr_receipt/gui/projects_tab.py
@@ -3,291 +3,18 @@ ProjectsTab: Project management interface for the OCR Invoice Parser.
 """
 from PyQt6.QtWidgets import (
     QWidget, QVBoxLayout, QHBoxLayout, QPushButton, QLabel, 
-    QTableWidget, QTableWidgetItem, QHeaderView, QMessageBox,
-    QLineEdit, QTextEdit, QDialog, QFormLayout, QDialogButtonBox,
-    QSplitter, QFrame
+    QMessageBox, QFrame, QDialog
 )
-from PyQt6.QtCore import Qt, pyqtSignal
-from PyQt6.QtGui import QFont, QIcon
-from typing import Optional, Dict, Any, List
+from PyQt6.QtCore import pyqtSignal
+from PyQt6.QtGui import QFont
+from typing import Optional, Dict, Any
 from ..business.project_manager import ProjectManager
 from ..utils.translation_helper import tr
+from .widgets.projects_table import ProjectsTable
+from .dialogs.add_project_dialog import AddProjectDialog, EditProjectDialog
 import logging
 
 logger = logging.getLogger(__name__)
-
-class AddProjectDialog(QDialog):
-    """Dialog for adding a new project."""
-    
-    def __init__(self, parent: Optional[QWidget] = None) -> None:
-        super().__init__(parent)
-        self.setWindowTitle(tr("add_project_dialog.title"))
-        self.setModal(True)
-        self.setMinimumWidth(400)
-        self.setMinimumHeight(200)
-        self.project_name: Optional[str] = None
-        self.project_description: Optional[str] = None
-        self._setup_ui()
-
-    def _setup_ui(self) -> None:
-        layout = QVBoxLayout(self)
-        
-        # Header
-        header_label = QLabel(tr("add_project_dialog.header"))
-        header_label.setStyleSheet("font-size: 14px; font-weight: bold; margin-bottom: 10px;")
-        layout.addWidget(header_label)
-        
-        # Description
-        desc_label = QLabel(tr("add_project_dialog.description"))
-        desc_label.setWordWrap(True)
-        desc_label.setStyleSheet("color: #666; margin-bottom: 10px;")
-        layout.addWidget(desc_label)
-        
-        # Form layout
-        form_layout = QFormLayout()
-        
-        # Project name input
-        self.name_edit = QLineEdit()
-        self.name_edit.setPlaceholderText(tr("add_project_dialog.name_placeholder"))
-        self.name_edit.setMaxLength(100)
-        form_layout.addRow(tr("add_project_dialog.name_label"), self.name_edit)
-        
-        # Project description input
-        self.description_edit = QTextEdit()
-        self.description_edit.setPlaceholderText(tr("add_project_dialog.description_placeholder"))
-        self.description_edit.setMaximumHeight(100)
-        form_layout.addRow(tr("add_project_dialog.description_label"), self.description_edit)
-        
-        layout.addLayout(form_layout)
-        
-        # Buttons
-        button_box = QDialogButtonBox(QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel)
-        button_box.accepted.connect(self._on_accept)
-        button_box.rejected.connect(self.reject)
-        layout.addWidget(button_box)
-        
-        # Connect signals
-        self.name_edit.textChanged.connect(self._on_text_changed)
-        
-        # Set initial state
-        self._on_text_changed()
-
-    def _on_text_changed(self) -> None:
-        """Handle text changes for validation."""
-        text = self.name_edit.text().strip()
-        ok_button = self.findChild(QDialogButtonBox).button(QDialogButtonBox.StandardButton.Ok)
-        ok_button.setEnabled(bool(text))
-
-    def _on_accept(self) -> None:
-        """Handle accept button click with validation."""
-        name = self.name_edit.text().strip()
-        description = self.description_edit.toPlainText().strip()
-        
-        # Validation
-        if not name:
-            QMessageBox.warning(self, tr("add_project_dialog.validation_error"), 
-                              tr("add_project_dialog.enter_name"))
-            self.name_edit.setFocus()
-            return
-        
-        if len(name) < 2:
-            QMessageBox.warning(self, tr("add_project_dialog.validation_error"), 
-                              tr("add_project_dialog.min_length"))
-            self.name_edit.setFocus()
-            return
-        
-        if len(name) > 100:
-            QMessageBox.warning(self, tr("add_project_dialog.validation_error"), 
-                              tr("add_project_dialog.max_length"))
-            self.name_edit.setFocus()
-            return
-        
-        self.project_name = name
-        self.project_description = description if description else None
-        self.accept()
-
-    def get_project_name(self) -> Optional[str]:
-        """Get the entered project name."""
-        return self.project_name
-
-    def get_project_description(self) -> Optional[str]:
-        """Get the entered project description."""
-        return self.project_description
-
-class EditProjectDialog(QDialog):
-    """Dialog for editing an existing project."""
-    
-    def __init__(self, project_data: Dict[str, Any], parent: Optional[QWidget] = None) -> None:
-        super().__init__(parent)
-        self.project_data = project_data
-        self.setWindowTitle(tr("edit_project_dialog.title"))
-        self.setModal(True)
-        self.setMinimumWidth(400)
-        self.setMinimumHeight(200)
-        self.project_name: Optional[str] = None
-        self.project_description: Optional[str] = None
-        self._setup_ui()
-
-    def _setup_ui(self) -> None:
-        layout = QVBoxLayout(self)
-        
-        # Header
-        header_label = QLabel(tr("edit_project_dialog.header"))
-        header_label.setStyleSheet("font-size: 14px; font-weight: bold; margin-bottom: 10px;")
-        layout.addWidget(header_label)
-        
-        # Form layout
-        form_layout = QFormLayout()
-        
-        # Project name input
-        self.name_edit = QLineEdit()
-        self.name_edit.setText(self.project_data.get('name', ''))
-        self.name_edit.setPlaceholderText(tr("edit_project_dialog.name_placeholder"))
-        self.name_edit.setMaxLength(100)
-        form_layout.addRow(tr("edit_project_dialog.name_label"), self.name_edit)
-        
-        # Project description input
-        self.description_edit = QTextEdit()
-        self.description_edit.setText(self.project_data.get('description', ''))
-        self.description_edit.setPlaceholderText(tr("edit_project_dialog.description_placeholder"))
-        self.description_edit.setMaximumHeight(100)
-        form_layout.addRow(tr("edit_project_dialog.description_label"), self.description_edit)
-        
-        layout.addLayout(form_layout)
-        
-        # Buttons
-        button_box = QDialogButtonBox(QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel)
-        button_box.accepted.connect(self._on_accept)
-        button_box.rejected.connect(self.reject)
-        layout.addWidget(button_box)
-        
-        # Connect signals
-        self.name_edit.textChanged.connect(self._on_text_changed)
-        
-        # Set initial state
-        self._on_text_changed()
-
-    def _on_text_changed(self) -> None:
-        """Handle text changes for validation."""
-        text = self.name_edit.text().strip()
-        ok_button = self.findChild(QDialogButtonBox).button(QDialogButtonBox.StandardButton.Ok)
-        ok_button.setEnabled(bool(text))
-
-    def _on_accept(self) -> None:
-        """Handle accept button click with validation."""
-        name = self.name_edit.text().strip()
-        description = self.description_edit.toPlainText().strip()
-        
-        # Validation
-        if not name:
-            QMessageBox.warning(self, tr("edit_project_dialog.validation_error"), 
-                              tr("edit_project_dialog.enter_name"))
-            self.name_edit.setFocus()
-            return
-        
-        if len(name) < 2:
-            QMessageBox.warning(self, tr("edit_project_dialog.validation_error"), 
-                              tr("edit_project_dialog.min_length"))
-            self.name_edit.setFocus()
-            return
-        
-        if len(name) > 100:
-            QMessageBox.warning(self, tr("edit_project_dialog.validation_error"), 
-                              tr("edit_project_dialog.max_length"))
-            self.name_edit.setFocus()
-            return
-        
-        self.project_name = name
-        self.project_description = description if description else None
-        self.accept()
-
-    def get_project_name(self) -> Optional[str]:
-        """Get the entered project name."""
-        return self.project_name
-
-    def get_project_description(self) -> Optional[str]:
-        """Get the entered project description."""
-        return self.project_description
-
-class ProjectsTable(QTableWidget):
-    """Table widget for displaying and managing projects."""
-    
-    project_selected = pyqtSignal(int)  # Emits project ID when a project is selected
-    project_double_clicked = pyqtSignal(int)  # Emits project ID when a project is double-clicked
-    
-    def __init__(self, parent: Optional[QWidget] = None) -> None:
-        super().__init__(parent)
-        self._setup_ui()
-        self.projects_data: List[Dict[str, Any]] = []
-
-    def _setup_ui(self) -> None:
-        """Setup the table UI."""
-        self.setColumnCount(2)
-        self.setHorizontalHeaderLabels([
-            tr("projects_table.name_header"),
-            tr("projects_table.description_header")
-        ])
-        
-        # Set column widths
-        header = self.horizontalHeader()
-        header.setSectionResizeMode(0, QHeaderView.ResizeMode.Stretch)
-        header.setSectionResizeMode(1, QHeaderView.ResizeMode.Stretch)
-        
-        # Set selection behavior
-        self.setSelectionBehavior(QTableWidget.SelectionBehavior.SelectRows)
-        self.setSelectionMode(QTableWidget.SelectionMode.SingleSelection)
-        
-        # Connect signals
-        self.itemSelectionChanged.connect(self._on_selection_changed)
-
-    def load_projects(self, projects: List[Dict[str, Any]]) -> None:
-        """Load projects into the table."""
-        self.projects_data = projects
-        self.setRowCount(len(projects))
-        
-        for row, project in enumerate(projects):
-            # Name column
-            name_item = QTableWidgetItem(project['name'])
-            name_item.setFlags(name_item.flags() & ~Qt.ItemFlag.ItemIsEditable)
-            self.setItem(row, 0, name_item)
-            
-            # Description column
-            description = project.get('description', '') or ''
-            desc_item = QTableWidgetItem(description)
-            desc_item.setFlags(desc_item.flags() & ~Qt.ItemFlag.ItemIsEditable)
-            self.setItem(row, 1, desc_item)
-
-    def get_selected_project_id(self) -> Optional[int]:
-        """Get the ID of the currently selected project."""
-        current_row = self.currentRow()
-        if current_row >= 0 and current_row < len(self.projects_data):
-            return self.projects_data[current_row]['id']
-        return None
-
-    def get_selected_project_data(self) -> Optional[Dict[str, Any]]:
-        """Get the data of the currently selected project."""
-        current_row = self.currentRow()
-        if current_row >= 0 and current_row < len(self.projects_data):
-            return self.projects_data[current_row]
-        return None
-
-    def _on_selection_changed(self) -> None:
-        """Handle selection changes."""
-        project_id = self.get_selected_project_id()
-        if project_id is not None:
-            self.project_selected.emit(project_id)
-
-    def mouseDoubleClickEvent(self, event) -> None:
-        """Handle double-click events on table rows."""
-        super().mouseDoubleClickEvent(event)
-        
-        # Get the item that was double-clicked
-        item = self.itemAt(event.pos())
-        if item is not None:
-            row = item.row()
-            if row >= 0 and row < len(self.projects_data):
-                project_id = self.projects_data[row]['id']
-                self.project_double_clicked.emit(project_id)
 
 class ProjectsTab(QWidget):
     """Main projects management tab."""
@@ -492,7 +219,6 @@ class ProjectsTab(QWidget):
         
         # Update table headers
         self.projects_table.setHorizontalHeaderLabels([
-            tr("projects_table.id_header"),
             tr("projects_table.name_header"),
             tr("projects_table.description_header")
         ])
